fix conn setup error
the problem was the place of "await send_message".

I changed that the pending connection info update
is before the "await send_message".

initiator or intermediate nodes store the request to
the pending_connections as "connection info". and then
the connection setup response sent from the responder
to the node and the node move the conn info from
the pending connection to the running connection.

if the initiator send the conn setup req message to the next node,
the responder receive it and reply it. at the moment, the initiator
waits the response of the first request. so the pending connection
doesn't contain the conn info.